+++ test/test-project/src/test/editor/editorView-test.ts
@@ -88,9 +88,20 @@ describe('EditorView', function () {
         expect(actions).not.empty;
     });
 
-    it('getAction works', async function () {
+    it('getAction(title: string) works', async function () {
         const action = await view.getAction('More Actions...');
-        expect(action).not.undefined;
+        expect(await action.getTitle()).equal('More Actions...');
+    });
+
+    it('getAction(predicate: PredicateFunction) works', async function () {
+        const action = await view.getAction(async (action) => await action.getTitle() === 'More Actions...');
+        expect(await action.getTitle()).equal('More Actions...');
+    });
+
+    it('Editor getAction works', async function () {
+        await new EditorView().openEditor('Untitled-2');
+        const editorAction = await view.getAction('Hello World');
+        expect(editorAction).not.undefined;
     });
 
     describe('Editor Tab', function () {